Implemented the new CPP to constructors

+++ app/helpers.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 function formatDate(string $date): string
 {
-    $date = date('M d, Y', strtotime($date));
+    $date = date('M d, Y', (int)strtotime($date));
 
     return $date;
 }
